Improve mismatched types diagnostic
When a trait `Tr` is implemented for `Never` and the trait solver
attempts to resolve `T: Tr`, the old implementation would eventually
rewrite the obligation to `fn() -> !: FnOnce<(), Output = T>`, which
fails with the following error:

	error[E0271]: expected `fn() -> !` to return `T`, but it returns `!`
	note: required for `fn() -> !` to implement `fn_traits::FnOnce<()>`
	    impl<F, R> FnOnce<()> for F
		       ^^^^^^^^^^     ^
	    where
		F: core::ops::FnOnce() -> R,
					  - unsatisfied trait bound introduced here

This is quite misleading, because `fn() -> !` *does* implement
`FnOnce<()>`, just not with the right type; and it differs from how all
other type mismatch errors are reported, which is:

	error[E0277]: the trait bound `T: Tr` is not satisfied
	help: the trait `Tr` is implemented for `!`

This commit improves the error message to something very close:

	error[E0277]: the trait bound `T: Tr` is not satisfied
	help: the trait `Tr` is implemented for `<fn() -> ! as FnPtr>::Output`

How often this matters for `never-say-never` is questionable, because
`impl Tr for Never` makes implementing the trait for any other type
impossible due to coherence. But as `never-say-never` is often refered
to as the canonical implementation of the hack, it's probably better to
keep it high-quality.